mvvm-client/requirements.txt
pyedifice==4.4.1
PyQt6==6.10.0